requirements.txt
lxml==4.6.4
Pillow==8.4.0
pycryptodome==3.11.0
qrcode==7.3.1
